Fix bugs

+++ src/main/java/com/example/rest/entities/RecyclingPoint.java
@@ -1,6 +1,5 @@
 package com.example.rest.entities;
 
-import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -9,8 +8,8 @@ import lombok.Data;
 public class RecyclingPoint {
     private int id;
     private String city;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
     private String specialization;
     private String organization;
     private String phone;
